Fix gcc 4.4 warnings.
Legacy-ID: 4125

+++ debugger/debugger_internals.h
@@ -58,7 +58,7 @@ void debugger_exit_emulator( void );
 
 /* Utility functions called by the flex scanner */
 
-int debugger_command_input( char *buf, size_t *result, int max_size );
+int debugger_command_input( char *buf, int *result, int max_size );
 int debugger_page_hash( const char *text );
 int yylex( void );
 void yyerror( const char *s );
